split common util
